refactor(cli): decompose agent-relay-mcp god file into mcp/ modules
Split the 2215-line packages/cli/src/cli/agent-relay-mcp.ts into eight
single-responsibility modules under packages/cli/src/cli/mcp/:

- types.ts          shared session/option types
- tool-results.ts   MCP tool-result formatting
- action-schema.ts  JSON-schema <-> zod adapter + descriptor serialization
- resources.ts      realtime resource bridge + relay:// resource registration
- inbox.ts          inbox summary formatting + piggyback skip set
- telemetry.ts      tool-call telemetry + inbox-piggyback registerTool wrapper
- workspace.ts      workspace provisioning helpers
- action-tools.ts   list_actions/invoke_action + dynamic per-action tools

Pure move with no behavior change; the public export surface is unchanged.
The orchestration file drops from 2215 to ~1390 lines.
