+++ packages/cli/src/cli/agent-relay-mcp.ts
@@ -4,39 +4,43 @@ import fs from 'node:fs';
 import path from 'node:path';
 import { fileURLToPath } from 'node:url';
 
-import { McpServer, ResourceTemplate } from '@modelcontextprotocol/sdk/server/mcp.js';
+import { McpServer } from '@modelcontextprotocol/sdk/server/mcp.js';
 import { StdioServerTransport } from '@modelcontextprotocol/sdk/server/stdio.js';
 import {
   ListToolsRequestSchema,
   SubscribeRequestSchema,
   UnsubscribeRequestSchema,
 } from '@modelcontextprotocol/sdk/types.js';
-import { RelayCast, SDK_VERSION, WsClient, type AgentClient } from '@relaycast/sdk';
-import {
-  INVALID_AGENT_TOKEN_CODE,
-  agentTokenRecoveryMessage,
-  isInvalidAgentTokenError,
-  isInvalidAgentTokenToolResult,
-} from '@agent-relay/sdk';
+import { RelayCast, SDK_VERSION, WsClient } from '@relaycast/sdk';
 import { AgentRelay } from '@agent-relay/sdk';
-import type {
-  ActionAuditEvent,
-  ActionSchema,
-  AgentRelayActionDescriptor,
-  AgentRelayActions,
-  JsonSchemaLiteObject,
-  ZodLikeSchema,
-} from '@agent-relay/sdk/actions';
 import { z } from 'zod';
+import { initTelemetry, shutdown as shutdownTelemetry } from './telemetry/index.js';
+import { withRelaycastTelemetry } from './lib/relaycast-telemetry.js';
 import {
-  initTelemetry,
-  shutdown as shutdownTelemetry,
-  track,
-  type AgentRelayToolCallCategory,
-  type AgentRelayToolCallType,
-} from './telemetry/index.js';
-import { relaycastWorkspaceTelemetryOptions, withRelaycastTelemetry } from './lib/relaycast-telemetry.js';
-import { errorClassName } from './lib/telemetry-helpers.js';
+  RealtimeResourceBridge,
+  SubscriptionManager,
+  registerResourceDefinitions,
+} from './mcp/resources.js';
+import { jsonContent, jsonResult, textContent } from './mcp/tool-results.js';
+import {
+  createWorkspace,
+  extractWorkspaceKey,
+  extractWorkspaceName,
+  requireWorkspaceKey,
+} from './mcp/workspace.js';
+import { enableInboxPiggyback } from './mcp/telemetry.js';
+import { registerAgentRelayActionTools } from './mcp/action-tools.js';
+import type {
+  AgentClientLike,
+  AgentRelayMcpServerOptions,
+  AgentType,
+  RegisteredAgent,
+  RegistrationSession,
+  RelayCastLike,
+  SessionSetter,
+  SessionState,
+} from './mcp/types.js';
+export type { AgentRelayMcpServerOptions } from './mcp/types.js';
 
 const DEFAULT_BASE_URL = 'https://gateway.relaycast.dev';
 export const AGENT_RELAY_MCP_VERSION = process.env.AGENT_RELAY_CLI_VERSION ?? SDK_VERSION ?? 'unknown';
@@ -77,7 +81,6 @@ const DEFAULT_SYSTEM_PROMPT = `You are an AI agent in a collaborative workspace 
 - React with emoji to acknowledge messages
 - Keep messages concise and actionable`;
 
-const jsonResult = z.object({}).passthrough();
 const messageResult = {
   message: z.string().describe('Human-readable confirmation message'),
 };
@@ -88,44 +91,6 @@ const identityOverrideInputShape = {
     .describe('Registered agent identity to act as when multiple identities have been registered'),
 };
 
-type AgentType = 'agent' | 'human';
-type RelayCastLike = Pick<RelayCast, 'agents'>;
-type AgentClientLike = AgentClient;
-
-export interface AgentRelayMcpServerOptions {
-  workspaceKey?: string;
-  /** @deprecated Use workspaceKey. */
-  apiKey?: string;
-  baseUrl?: string;
-  agentToken?: string;
-  agentName?: string;
-  agentType?: AgentType;
-  strictAgentName?: boolean;
-  telemetryTransport?: 'stdio' | 'http';
-  skipBootstrap?: boolean;
-  actions?: AgentRelayActions;
-  onActionAuditEvent?: (event: ActionAuditEvent) => Promise<void> | void;
-}
-
-interface RegisteredAgent {
-  agentName: string;
-  agentToken: string;
-}
-
-interface SessionState {
-  workspaceKey: string | null;
-  agentToken: string | null;
-  agentName: string | null;
-  agents: Map<string, RegisteredAgent>;
-  wsBridge: RealtimeResourceBridge | null;
-  subscriptions: SubscriptionManager | null;
-  wsInitAttempted: boolean;
-}
-
-type RegistrationSession = Pick<SessionState, 'workspaceKey' | 'agentToken' | 'agentName'> & {
-  agents?: Map<string, RegisteredAgent>;
-};
-type SessionSetter = (partial: Partial<SessionState>) => void;
 type AgentResultCallbackConfig = {
   url: string;
   token: string;
@@ -329,795 +294,6 @@ function createInitialSession(options: {
     subscriptions: null,
     wsInitAttempted: false,
   };
-}
-
-class SubscriptionManager {
-  private readonly subscriptions = new Set<string>();
-
-  subscribe(uri: string): void {
-    this.subscriptions.add(uri);
-  }
-
-  unsubscribe(uri: string): void {
-    this.subscriptions.delete(uri);
-  }
-
-  getMatchingSubscriptions(uris: string[]): string[] {
-    return uris.filter((uri) => this.subscriptions.has(uri));
-  }
-
-  getAll(): string[] {
-    return [...this.subscriptions];
-  }
-
-  clear(): void {
-    this.subscriptions.clear();
-  }
-}
-
-function getStringEventField(event: unknown, field: string): string | null {
-  if (typeof event !== 'object' || event === null) {
-    return null;
-  }
-  const candidate = (event as Record<string, unknown>)[field];
-  return typeof candidate === 'string' ? candidate : null;
-}
-
-function eventToResourceUris(event: unknown): string[] {
-  const type = getStringEventField(event, 'type');
-  switch (type) {
-    case 'message.created': {
-      const channel = getStringEventField(event, 'channel');
-      return channel ? ['relay://inbox', `relay://channels/${channel}/messages`] : ['relay://inbox'];
-    }
-    case 'message.updated': {
-      const channel = getStringEventField(event, 'channel');
-      return channel ? [`relay://channels/${channel}/messages`] : [];
-    }
-    case 'thread.reply': {
-      const parentId = getStringEventField(event, 'parentId');
-      return parentId ? ['relay://inbox', `relay://messages/${parentId}/thread`] : ['relay://inbox'];
-    }
-    case 'dm.received':
-    case 'group_dm.received': {
-      const conversationId = getStringEventField(event, 'conversationId');
-      return conversationId ? ['relay://inbox', `relay://dm/${conversationId}`] : ['relay://inbox'];
-    }
-    case 'agent.online':
-    case 'agent.offline':
-      return ['relay://agents'];
-    case 'channel.created':
-    case 'channel.updated':
-    case 'channel.archived':
-    case 'member.joined':
-    case 'member.left':
-      return ['relay://channels'];
-    case 'webhook.received':
-    case 'command.invoked': {
-      const channel = getStringEventField(event, 'channel');
-      return channel ? [`relay://channels/${channel}/messages`] : [];
-    }
-    case 'reaction.added':
-    case 'reaction.removed':
-      return ['relay://inbox'];
-    default:
-      return [];
-  }
-}
-
-class RealtimeResourceBridge {
-  private unsubscribeFn: (() => void) | null = null;
-
-  constructor(
-    private readonly wsClient: WsClient,
-    private readonly subscriptions: SubscriptionManager,
-    private readonly notifyCallback: (uri: string) => void
-  ) {}
-
-  start(): void {
-    this.unsubscribeFn = this.wsClient.on('*', (event) => {
-      const type = getStringEventField(event, 'type');
-      if (
-        type === 'open' ||
-        type === 'close' ||
-        type === 'error' ||
-        type === 'reconnecting' ||
-        type === 'permanently_disconnected'
-      ) {
-        return;
-      }
-      const matched = this.subscriptions.getMatchingSubscriptions(eventToResourceUris(event));
-      for (const uri of matched) {
-        this.notifyCallback(uri);
-      }
-    });
-    this.wsClient.connect();
-  }
-
-  stop(): void {
-    if (this.unsubscribeFn) {
-      this.unsubscribeFn();
-      this.unsubscribeFn = null;
-    }
-    this.wsClient.disconnect();
-  }
-}
-
-function registerResourceDefinitions(
-  server: McpServer,
-  getAgentClient: (asIdentity?: string) => AgentClientLike,
-  getRelay: () => RelayCast
-): void {
-  server.registerResource(
-    'inbox',
-    'relay://inbox',
-    { title: 'Inbox', description: 'Unread messages, mentions, and DMs', mimeType: 'application/json' },
-    async (uri) => {
-      const inbox = await getAgentClient().inbox();
-      return { contents: [{ uri: uri.href, text: JSON.stringify(inbox) }] };
-    }
-  );
-
-  server.registerResource(
-    'agents',
-    'relay://agents',
-    {
-      title: 'Agents',
-      description: 'Online and offline agents in the workspace',
-      mimeType: 'application/json',
-    },
-    async (uri) => {
-      const agents = await getRelay().agents.list();
-      return { contents: [{ uri: uri.href, text: JSON.stringify(agents) }] };
-    }
-  );
-
-  server.registerResource(
-    'channels',
-    'relay://channels',
-    { title: 'Channels', description: 'Available channels in the workspace', mimeType: 'application/json' },
-    async (uri) => {
-      const channels = await getAgentClient().channels.list();
-      return { contents: [{ uri: uri.href, text: JSON.stringify(channels) }] };
-    }
-  );
-
-  server.registerResource(
-    'channel-messages',
-    new ResourceTemplate('relay://channels/{name}/messages', { list: undefined }),
-    {
-      title: 'Channel Messages',
-      description: 'Messages in a specific channel',
-      mimeType: 'application/json',
-    },
-    async (uri, params) => {
-      const messages = await getAgentClient().messages(String(params.name));
-      return { contents: [{ uri: uri.href, text: JSON.stringify(messages) }] };
-    }
-  );
-
-  server.registerResource(
-    'message-thread',
-    new ResourceTemplate('relay://messages/{id}/thread', { list: undefined }),
-    { title: 'Message Thread', description: 'Thread replies on a message', mimeType: 'application/json' },
-    async (uri, params) => {
-      const thread = await getAgentClient().thread(String(params.id));
-      return { contents: [{ uri: uri.href, text: JSON.stringify(thread) }] };
-    }
-  );
-
-  server.registerResource(
-    'dm-conversation',
-    new ResourceTemplate('relay://dm/{conversation_id}', { list: undefined }),
-    {
-      title: 'DM Conversation',
-      description: 'Direct message conversation',
-      mimeType: 'application/json',
-    },
-    async (uri, params) => {
-      const messages = await getAgentClient().dms.messages(String(params.conversation_id));
-      return { contents: [{ uri: uri.href, text: JSON.stringify(messages) }] };
-    }
-  );
-}
-
-function hasContentArray(value: unknown): value is { content: Array<Record<string, unknown>> } {
-  return (
-    typeof value === 'object' && value !== null && Array.isArray((value as { content?: unknown }).content)
-  );
-}
-
-const SKIP_PIGGYBACK = new Set(['check_inbox', 'create_workspace', 'set_workspace_key', 'register_agent']);
-
-function formatInbox(inbox: any, selfName?: string | null): string {
-  const norm = (s: string) => s.trim().replace(/^@/, '').toLowerCase();
-  const selfNorm = selfName ? norm(selfName) : null;
-  const isSelf = (name: string) => selfNorm != null && norm(name) === selfNorm;
-  const lines = ['--- Pending Messages ---'];
-
-  if (inbox.unreadChannels?.length) {
-    lines.push('Unread channels:');
-    for (const ch of inbox.unreadChannels) {
-      lines.push(`  #${ch.channelName}: ${ch.unreadCount} unread`);
-    }
-  }
-
-  const mentions = selfNorm ? inbox.mentions?.filter((m: any) => !isSelf(m.agentName)) : inbox.mentions;
-  if (mentions?.length) {
-    lines.push('Mentions:');
-    for (const m of mentions) {
-      lines.push(`  @${m.agentName} in #${m.channelName}: "${m.text}"`);
-    }
-  }
-
-  const dms = selfNorm ? inbox.unreadDms?.filter((dm: any) => !isSelf(dm.from)) : inbox.unreadDms;
-  if (dms?.length) {
-    lines.push('Unread DMs:');
-    for (const dm of dms) {
-      lines.push(`  From ${dm.from}: ${dm.unreadCount} unread`);
-    }
-  }
-
-  const reactions = selfNorm
-    ? inbox.recentReactions?.filter((reaction: any) => !isSelf(reaction.agentName))
-    : inbox.recentReactions;
-  if (reactions?.length) {
-    lines.push('Reactions (informational; no response required):');
-    for (const reaction of reactions) {
-      lines.push(
-        `  :${reaction.emoji}: on your message in #${reaction.channelName} by @${reaction.agentName}`
-      );
-    }
-  }
-
-  return lines.length === 1 ? '' : lines.join('\n');
-}
-
-function readAsIdentity(args: unknown[]): string | undefined {
-  const [input] = args;
-  if (typeof input !== 'object' || input === null) return undefined;
-  const as = (input as { as?: unknown }).as;
-  return typeof as === 'string' ? as : undefined;
-}
-
-function invalidAgentTokenToolResult(): JsonToolResult & { isError: true } {
-  const text = agentTokenRecoveryMessage();
-  return {
-    content: [{ type: 'text', text }],
-    structuredContent: {
-      error: { code: INVALID_AGENT_TOKEN_CODE, message: text },
-    },
-    isError: true,
-  };
-}
-
-function isErrorToolResult(value: unknown): boolean {
-  return Boolean(value && typeof value === 'object' && (value as { isError?: unknown }).isError === true);
-}
-
-interface AgentRelayToolCallMetadata {
-  toolType: AgentRelayToolCallType;
-  toolCategory: AgentRelayToolCallCategory;
-}
-
-/**
- * Owned tools that delegate to the actions surface (`actions.invoke(...)`)
- * rather than the agents/messaging APIs. Together with the dynamic per-action
- * tools (tracked via `actionToolNames`), these intentionally skip per-tool
- * telemetry so the same underlying action is not counted differently depending
- * on which MCP surface the caller used (e.g. `spawn` vs `invoke_action`).
- */
-const ACTION_ROUTED_TOOL_NAMES = new Set(['invoke_action', 'spawn']);
-
-/**
- * Coarse type/category metadata for the statically-registered ("owned") MCP
- * tools. Action-routed calls (see `ACTION_ROUTED_TOOL_NAMES`) and the dynamic
- * per-action tools surfaced from the actions registry are intentionally
- * excluded from per-tool telemetry (see the skip in `enableInboxPiggyback`),
- * so they have no entry.
- */
-const AGENT_RELAY_TOOL_CALL_METADATA = {
-  add_agent: { toolType: 'agent.create', toolCategory: 'spawn' },
-  remove_agent: { toolType: 'agent.release', toolCategory: 'release' },
-  list_actions: { toolType: 'action.list', toolCategory: 'action' },
-  submit_result: { toolType: 'result.submit', toolCategory: 'result' },
-  create_workspace: { toolType: 'workspace.create', toolCategory: 'workspace' },
-  set_workspace_key: { toolType: 'workspace.set_key', toolCategory: 'workspace' },
-  register_agent: { toolType: 'agent.register', toolCategory: 'agent' },
-  list_agents: { toolType: 'agent.list', toolCategory: 'agent' },
-  post_message: { toolType: 'message.post', toolCategory: 'message' },
-  send_dm: { toolType: 'message.dm', toolCategory: 'message' },
-  send_group_dm: { toolType: 'message.group_dm', toolCategory: 'message' },
-  list_dms: { toolType: 'message.dm_list', toolCategory: 'message' },
-  list_messages: { toolType: 'message.list', toolCategory: 'message' },
-  reply_to_thread: { toolType: 'message.reply', toolCategory: 'message' },
-  get_message_thread: { toolType: 'message.thread', toolCategory: 'message' },
-  search_messages: { toolType: 'message.search', toolCategory: 'message' },
-  create_channel: { toolType: 'channel.create', toolCategory: 'channel' },
-  list_channels: { toolType: 'channel.list', toolCategory: 'channel' },
-  join_channel: { toolType: 'channel.join', toolCategory: 'channel' },
-  leave_channel: { toolType: 'channel.leave', toolCategory: 'channel' },
-  set_channel_topic: { toolType: 'channel.set_topic', toolCategory: 'channel' },
-  archive_channel: { toolType: 'channel.archive', toolCategory: 'channel' },
-  invite_to_channel: { toolType: 'channel.invite', toolCategory: 'channel' },
-  add_reaction: { toolType: 'reaction.add', toolCategory: 'reaction' },
-  remove_reaction: { toolType: 'reaction.remove', toolCategory: 'reaction' },
-  check_inbox: { toolType: 'inbox.check', toolCategory: 'inbox' },
-  mark_message_read: { toolType: 'inbox.mark_read', toolCategory: 'inbox' },
-  get_message_readers: { toolType: 'inbox.reader_list', toolCategory: 'inbox' },
-} satisfies Record<string, AgentRelayToolCallMetadata>;
-
-function agentRelayToolCallMetadata(name: string): AgentRelayToolCallMetadata {
-  const known = (AGENT_RELAY_TOOL_CALL_METADATA as Partial<Record<string, AgentRelayToolCallMetadata>>)[name];
-  return known ?? { toolType: name, toolCategory: 'tool' };
-}
-
-function trackAgentRelayToolCall(input: {
-  toolName: string;
-  toolType: AgentRelayToolCallType;
-  toolCategory: AgentRelayToolCallCategory;
-  transport?: AgentRelayMcpServerOptions['telemetryTransport'];
-  startedAt: number;
-  success: boolean;
-  errorClass?: string;
-}): void {
-  track('agent_relay_tool_call', {
-    tool_name: input.toolName,
-    tool_type: input.toolType,
-    tool_category: input.toolCategory,
-    transport: input.transport ?? 'unknown',
-    success: input.success,
-    duration_ms: Date.now() - input.startedAt,
-    ...(input.errorClass ? { error_class: input.errorClass } : {}),
-  });
-}
-
-function enableInboxPiggyback(
-  mcpServer: McpServer,
-  getSession: () => SessionState,
-  getAgentClient: (asIdentity?: string) => AgentClientLike,
-  invalidateAgentToken: (asIdentity?: string) => void,
-  telemetryTransport?: AgentRelayMcpServerOptions['telemetryTransport'],
-  actionToolNames = new Set<string>()
-): void {
-  const original = mcpServer.registerTool.bind(mcpServer);
-  const mutableServer = mcpServer as McpServer & {
-    registerTool: McpServer['registerTool'];
-  };
-
-  mutableServer.registerTool = (name: string, config: any, handler: any) => {
-    if (!handler) {
-      return original(name, config, handler);
-    }
-
-    const wrapped = async (...args: unknown[]) => {
-      const asIdentity = readAsIdentity(args);
-      const startedAt = Date.now();
-      // Action-routed calls (`invoke_action`, `spawn`, and the dynamic
-      // per-action tools) run through the actions surface and deliberately skip
-      // per-tool telemetry; only the owned tools emit `agent_relay_tool_call`.
-      const toolMetadata =
-        !ACTION_ROUTED_TOOL_NAMES.has(name) && !actionToolNames.has(name)
-          ? agentRelayToolCallMetadata(name)
-          : undefined;
-
-      let result: any;
-      try {
-        result = await handler(...args);
-      } catch (err) {
-        if (name !== 'register_agent' && isInvalidAgentTokenError(err)) {
-          invalidateAgentToken(asIdentity);
-          if (toolMetadata) {
-            trackAgentRelayToolCall({
-              toolName: name,
-              toolType: toolMetadata.toolType,
-              toolCategory: toolMetadata.toolCategory,
-              transport: telemetryTransport,
-              startedAt,
-              success: false,
-              errorClass: errorClassName(err) ?? 'InvalidAgentToken',
-            });
-          }
-          return invalidAgentTokenToolResult();
-        }
-        if (toolMetadata) {
-          trackAgentRelayToolCall({
-            toolName: name,
-            toolType: toolMetadata.toolType,
-            toolCategory: toolMetadata.toolCategory,
-            transport: telemetryTransport,
-            startedAt,
-            success: false,
-            errorClass: errorClassName(err),
-          });
-        }
-        throw err;
-      }
-
-      if (name !== 'register_agent' && isInvalidAgentTokenToolResult(result)) {
-        invalidateAgentToken(asIdentity);
-        if (toolMetadata) {
-          trackAgentRelayToolCall({
-            toolName: name,
-            toolType: toolMetadata.toolType,
-            toolCategory: toolMetadata.toolCategory,
-            transport: telemetryTransport,
-            startedAt,
-            success: false,
-            errorClass: 'InvalidAgentToken',
-          });
-        }
-        if (hasContentArray(result)) {
-          result.content.push({ type: 'text', text: agentTokenRecoveryMessage() });
-        }
-        return result;
-      }
-
-      if (!SKIP_PIGGYBACK.has(name) && getSession().agentToken && hasContentArray(result)) {
-        try {
-          const inbox = await getAgentClient(asIdentity).inbox();
-          const inboxText = formatInbox(inbox, asIdentity ?? getSession().agentName);
-          if (inboxText) {
-            result.content.push({ type: 'text', text: inboxText });
-          }
-        } catch (err) {
-          if (isInvalidAgentTokenError(err)) {
-            invalidateAgentToken(asIdentity);
-          }
-        }
-      }
-
-      if (toolMetadata) {
-        const resultIsError = isErrorToolResult(result);
-        trackAgentRelayToolCall({
-          toolName: name,
-          toolType: toolMetadata.toolType,
-          toolCategory: toolMetadata.toolCategory,
-          transport: telemetryTransport,
-          startedAt,
-          success: !resultIsError,
-          ...(resultIsError ? { errorClass: 'ToolResultError' } : {}),
-        });
-      }
-
-      return result;
-    };
-
-    return original(name, config, wrapped);
-  };
-}
-
-async function createWorkspace(name: string, baseUrl?: string): Promise<Record<string, unknown>> {
-  return (await RelayCast.createWorkspace(name, {
-    baseUrl,
-    ...relaycastWorkspaceTelemetryOptions(),
-  })) as Record<string, unknown>;
-}
-
-function extractWorkspaceKey(payload: Record<string, unknown>): string | undefined {
-  const data =
-    payload.data && typeof payload.data === 'object' ? (payload.data as Record<string, unknown>) : {};
-  const value =
-    payload.workspaceKey ??
-    payload.workspace_key ??
-    payload.apiKey ??
-    payload.api_key ??
-    data.workspaceKey ??
-    data.workspace_key ??
-    data.apiKey ??
-    data.api_key;
-
-  return typeof value === 'string' && value.trim() ? value : undefined;
-}
-
-function extractWorkspaceName(payload: Record<string, unknown>, fallback: string): string {
-  const data =
-    payload.data && typeof payload.data === 'object' ? (payload.data as Record<string, unknown>) : {};
-  const value = payload.workspaceName ?? payload.workspace_name ?? payload.name ?? data.workspaceName;
-  return typeof value === 'string' && value.trim() ? value : fallback;
-}
-
-function requireWorkspaceKey(session: RegistrationSession): void {
-  if (session.workspaceKey) {
-    return;
-  }
-
-  throw new Error(
-    'Workspace key not configured. Call "create_workspace" first, or "set_workspace_key" if someone shared a workspace key.'
-  );
-}
-
-type JsonToolResult = {
-  content: Array<{ type: 'text'; text: string }>;
-  structuredContent: Record<string, unknown>;
-};
-
-function jsonContent(value: unknown): JsonToolResult {
-  const structuredContent =
-    typeof value === 'object' && value !== null && !Array.isArray(value)
-      ? (value as Record<string, unknown>)
-      : { value };
-  return {
-    content: [{ type: 'text', text: JSON.stringify(value, null, 2) }],
-    structuredContent,
-  };
-}
-
-function textContent(message: string, structuredContent: Record<string, unknown> = { message }) {
-  return {
-    content: [{ type: 'text' as const, text: message }],
-    structuredContent,
-  };
-}
-
-function isSchemaObject(schema: ActionSchema | undefined): schema is JsonSchemaLiteObject {
-  return Boolean(
-    schema &&
-    typeof schema === 'object' &&
-    !Array.isArray(schema) &&
-    typeof (schema as { safeParse?: unknown }).safeParse !== 'function'
-  );
-}
-
-function getSchemaDescription(schema: ActionSchema | undefined): string | undefined {
-  return isSchemaObject(schema) && typeof schema.description === 'string' ? schema.description : undefined;
-}
-
-function zodFromJsonSchema(schema: ActionSchema | undefined): z.ZodTypeAny {
-  if (schema === false) {
-    return z.never();
-  }
-
-  if (!isSchemaObject(schema)) {
-    return z.unknown();
-  }
-
-  let zodType: z.ZodTypeAny;
-  const schemaType = Array.isArray(schema.type) ? schema.type[0] : schema.type;
-  switch (schemaType) {
-    case 'array':
-      zodType = z.array(zodFromJsonSchema(schema.items));
-      break;
-    case 'boolean':
-      zodType = z.boolean();
-      break;
-    case 'integer':
-      zodType = z.number().int();
-      break;
-    case 'number':
-      zodType = z.number();
-      break;
-    case 'object':
-      if (schema.properties) {
-        const required = new Set(schema.required ?? []);
-        const shape: Record<string, z.ZodTypeAny> = {};
-        for (const [key, childSchema] of Object.entries(schema.properties)) {
-          const child = zodFromJsonSchema(childSchema);
-          shape[key] = required.has(key) ? child : child.optional();
-        }
-        zodType = z.object(shape).passthrough();
-      } else {
-        zodType = z.record(z.string(), z.unknown());
-      }
-      break;
-    case 'string':
-      zodType = z.string();
-      break;
-    default:
-      zodType = z.unknown();
-      break;
-  }
-
-  const description = getSchemaDescription(schema);
-  return description ? zodType.describe(description) : zodType;
-}
-
-function actionToolInputSchema(schema: ActionSchema | undefined): Record<string, z.ZodTypeAny> {
-  const zodShape = zodObjectShape(schema);
-  if (zodShape) {
-    return zodShape;
-  }
-
-  if (!isSchemaObject(schema) || schema.type !== 'object') {
-    return {
-      input: z.unknown().describe('Action input payload. The action registry performs final validation.'),
-    };
-  }
-
-  const required = new Set(schema.required ?? []);
-  const shape: Record<string, z.ZodTypeAny> = {};
-  for (const [key, childSchema] of Object.entries(schema.properties ?? {})) {
-    const child = zodFromJsonSchema(childSchema);
-    shape[key] = required.has(key) ? child : child.optional();
-  }
-  return shape;
-}
-
-function actionInvocationInput(descriptor: AgentRelayActionDescriptor, args: unknown): unknown {
-  const schema = descriptor.inputSchema;
-  if (zodObjectShape(schema)) {
-    return args;
-  }
-  if (!isSchemaObject(schema) || schema.type !== 'object') {
-    return typeof args === 'object' && args !== null && 'input' in args
-      ? (args as { input?: unknown }).input
-      : args;
-  }
-  return args;
-}
-
-function zodObjectShape(schema: ActionSchema | undefined): Record<string, z.ZodTypeAny> | undefined {
-  if (schema instanceof z.ZodObject) {
-    return schema.shape;
-  }
-  return undefined;
-}
-
-function serializableActionDescriptor(descriptor: AgentRelayActionDescriptor): Record<string, unknown> {
-  return {
-    name: descriptor.name,
-    description: descriptor.description,
-    visibility: descriptor.visibility,
-    ...(descriptor.inputSchema ? { inputSchema: serializableActionSchema(descriptor.inputSchema) } : {}),
-    ...(descriptor.outputSchema ? { outputSchema: serializableActionSchema(descriptor.outputSchema) } : {}),
-  };
-}
-
-function serializableActionSchema(schema: ActionSchema): unknown {
-  if (isSchemaObject(schema)) {
-    return schema;
-  }
-  if (isZodLikeSchema(schema)) {
-    return {
-      type: 'zod',
-      ...(schema.description ? { description: schema.description } : {}),
-    };
-  }
-  return schema;
-}
-
-function isZodLikeSchema(schema: ActionSchema | undefined): schema is ZodLikeSchema {
-  return Boolean(
-    schema &&
-    typeof schema === 'object' &&
-    !Array.isArray(schema) &&
-    typeof (schema as { safeParse?: unknown }).safeParse === 'function'
-  );
-}
-
-function registerAgentRelayActionTools(
-  server: McpServer,
-  actions: AgentRelayActions | undefined,
-  getSession: () => SessionState,
-  onAuditEvent?: (event: ActionAuditEvent) => Promise<void> | void,
-  getAgentClient?: (asIdentity?: string) => AgentClientLike,
-  actionToolNames?: Set<string>
-): void {
-  if (!actions) {
-    return;
-  }
-
-  /**
-   * Fire-and-forget invocation through the relay: returns an immediate ack
-   * (with an `invocation_id`) and does NOT run the handler inline. Falls back to
-   * the local in-process registry when the relay action surface is unavailable.
-   */
-  const invokeAction = async (name: string, input: unknown) => {
-    const relayActions = getRelayAgentActions(getAgentClient);
-    if (relayActions) {
-      try {
-        const ack = await relayActions.invoke(name, asInputRecord(input));
-        return jsonContent({ ok: true, status: 'invoked', invocation: ack });
-      } catch (error) {
-        return { ...jsonContent({ ok: false, error: errorMessage(error) }), isError: true };
-      }
-    }
-
-    const session = getSession();
-    const result = await actions.invoke({
-      name,
-      input,
-      context: {
-        caller: { name: session.agentName ?? 'mcp', type: 'agent' },
-        emit: onAuditEvent,
-      },
-    });
-    return result.ok ? jsonContent(result) : { ...jsonContent(result), isError: true };
-  };
-
-  server.registerTool(
-    'list_actions',
-    {
-      title: 'List Actions',
-      description: 'List Agent Relay actions available to this agent.',
-      inputSchema: {},
-      outputSchema: jsonResult,
-      annotations: {
-        readOnlyHint: true,
-        destructiveHint: false,
-        idempotentHint: true,
-        openWorldHint: false,
-      },
-    },
-    async () =>
-      jsonContent({
-        actions: (await actions.list({ visibility: 'agent' })).map(serializableActionDescriptor),
-      })
-  );
-
-  server.registerTool(
-    'invoke_action',
-    {
-      title: 'Invoke Action',
-      description:
-        'Invoke a registered Agent Relay action by name. Fire-and-forget: returns an ack with an invocation id; the result arrives asynchronously to the action handler.',
-      inputSchema: {
-        name: z.string().describe('Registered action name'),
-        input: z.unknown().describe('Action input payload'),
-      },
-      outputSchema: jsonResult,
-      annotations: {
-        readOnlyHint: false,
-        destructiveHint: false,
-        idempotentHint: false,
-        openWorldHint: false,
-      },
-    },
-    async ({ name, input }: { name: string; input: unknown }) => invokeAction(name, input)
-  );
-
-  void actions
-    .list({ visibility: 'agent' })
-    .then((descriptors) => {
-      for (const descriptor of descriptors) {
-        actionToolNames?.add(descriptor.name);
-        server.registerTool(
-          descriptor.name,
-          {
-            title: descriptor.name,
-            description: descriptor.description,
-            inputSchema: actionToolInputSchema(descriptor.inputSchema),
-            outputSchema: jsonResult,
-            annotations: {
-              readOnlyHint: false,
-              destructiveHint: false,
-              idempotentHint: false,
-              openWorldHint: false,
-            },
-          },
-          async (args: unknown) => invokeAction(descriptor.name, actionInvocationInput(descriptor, args))
-        );
-      }
-    })
-    .catch(() => undefined);
-}
-
-/** The relay-backed action surface on the live agent client, when available. */
-function getRelayAgentActions(
-  getAgentClient?: (asIdentity?: string) => AgentClientLike
-): AgentClientLike['actions'] | undefined {
-  if (!getAgentClient) {
-    return undefined;
-  }
-  try {
-    return getAgentClient().actions;
-  } catch {
-    return undefined;
-  }
-}
-
-function asInputRecord(input: unknown): Record<string, unknown> | undefined {
-  if (input === undefined || input === null) {
-    return undefined;
-  }
-  if (typeof input === 'object' && !Array.isArray(input)) {
-    return input as Record<string, unknown>;
-  }
-  return { input };
-}
-
-function errorMessage(error: unknown): string {
-  return error instanceof Error ? error.message : String(error);
 }
 
 function createRegisteredAgent(agentName: string, agentToken: string): RegisteredAgent {
